fix: corrige panics de tipo uint/uint64 e troca login de funcionario para CPF
- Corrige interface conversion panic em historicoController.go e transacaoController.go (userID.(uint) -> userID.(uint64), remanescente da migracao do JWT)
- Adiciona Preload(Curso) em AdicionarSaldo por consistencia
- Adiciona campo CPF (unico, 11 digitos) em funcionario, substituindo telefone como identificador de login
- Atualiza schema.sql e seed de funcionario com CPFs de teste

+++ backend/controller/historicoController.go
@@ -15,7 +15,7 @@ func GetHistoricoAluno(c *gin.Context) {
 		return
 	}
 
-	rm := int(userID.(uint))
+	rm := int64(userID.(uint64))
 
 	var historico []model.Historico
 	if err := config.DB.Where("aluno_rm = ?", rm).Order("data_hora DESC").Find(&historico).Error; err != nil {
